[fix] security issue fix
글 작성시 세션을 활용하지 않던 문제 해결

+++ Namukkun/src/main/java/com/pard/namukkun/post/controller/PostController.java
@@ -25,50 +25,69 @@ public class PostController {
 
     @PostMapping("/upload/post")
     @Operation(summary = "게시물 등록", description = "게시물 내용을 입력해서 게시물을 등록합니다.")
-    public ResponseEntity<?> createPost(@RequestBody() PostCreateDTO postCreateDTO) {
+    public ResponseEntity<?> createPost(
+            @RequestBody() PostCreateDTO postCreateDTO,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한 확인 - 로그인 되어있으면 가능함
+        if (userId == null)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         // Post에서 첨부파일을 제외한 데이터는 CreateDTO형태로 받고, 첨부파일은 List형태로 따로 받음
         return postService.createPost(postCreateDTO);
     }
 
-    //-----------------------------------
-    @PostMapping("/test")
-    public ResponseEntity<?> test(){
-        log.info("test");
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-    //-----------------------------------
-
     @PostMapping("/upload/temppost")
     @Operation(summary = "게시물 임시저장", description = "게시물을 임시저장합니다.")
-    public ResponseEntity<?> createTempPost(@RequestBody() PostCreateDTO postCreateDTO) {
+    public ResponseEntity<?> createTempPost(
+            @RequestBody() PostCreateDTO postCreateDTO,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한 확인 - 로그인 되어있으면 가능함
+        if (userId == null)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         return postService.saveTempPost(postCreateDTO);
     }
 
     @PostMapping(value = "/upload/file", consumes = {"multipart/form-data"})
     @Operation(summary = "첨부파일 첨부", description = "첨부파일을 첨부합니다.")
-    public ReturnFileNameDTO uploadFile(@RequestPart("files") List<MultipartFile> files) {
-        return postService.uploadAttachment(files);
-    }
+    public ResponseEntity<?> uploadFile(
+            @SessionAttribute(name = "userid", required = false) Long userid,
+            @RequestPart("files") List<MultipartFile> files
+    ) {
+        //권한 확인 - 로그인 되어있으면 가능함
+        if (userid == null)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 
-//    // ------------ 유저 테스트용 ---------------
-//    @PostMapping(value = "/upload/img", consumes = {"multipart/form-data"})
-//    @Operation(summary = "이미지 첨부", description = "이미지 이름을 받아서 UUID를 앞에 붙인 이름을 반환합니다.")
-//    public ResponseEntity<?> uploadImg(@RequestParam("img") MultipartFile img){
-//        return postService.uploadImg(img);
-//    }
-//    // ------------ 유저 테스트용 ---------------
+        ReturnFileNameDTO dto = postService.uploadAttachment(files);
+        return new ResponseEntity<>(dto, HttpStatus.OK);
+    }
 
     @PostMapping(value = "/upload/img", consumes = {"multipart/form-data"})
     @Operation(summary = "이미지 첨부", description = "이미지 이름을 받아서 UUID를 앞에 붙인 이름을 반환합니다.")
-    public ResponseEntity<?> uploadImg(@RequestParam("img") MultipartFile img,
-                                       @SessionAttribute(name = "userid", required = false) Long userId){
+    public ResponseEntity<?> uploadImg(
+            @RequestParam("img") MultipartFile img,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한확인 - 로그인 되어있으면 가능함
+        if (userId == null)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         log.info("{}\n{}", userId, img);
-        return postService.uploadImg(img,userId);
+        return postService.uploadImg(img, userId);
     }
 
     @PostMapping("/delete/file")
     @Operation(summary = "첨부파일 삭제", description = "첨부파일 이름을 받아서 삭제합니다.")
-    public ResponseEntity<?> deleteFile(@RequestParam("fileName") String fileName) {
+    public ResponseEntity<?> deleteFile(
+            @RequestParam("fileName") String fileName,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한확인 - 로그인 되어있으면 가능함
+        if (userId == null)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         return s3AttachmentService.deleteByName(fileName);
     }
 
@@ -80,8 +99,16 @@ public class PostController {
 
     @GetMapping("/read/update")
     @Operation(summary = "수정할 게시물을 id를 통해서 읽습니다.")
-    public PostUpdateDTO findPostById(@RequestParam("id") Long id) {
-        return postService.findPostByIdUpdateVer(id);
+    public ResponseEntity<?> findPostById(
+            @RequestParam("id") Long id,  // 게시물 어이디
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한확인 - 글 작성자만 가능함
+        if (userId == null || !userId.equals(postService.getWriterUserId(id)))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        PostUpdateDTO dto = postService.findPostByIdUpdateVer(id);
+        return new ResponseEntity<>(dto, HttpStatus.OK);
     }
 
     @GetMapping("/read")
@@ -106,34 +133,63 @@ public class PostController {
 
     @GetMapping("/read/by-Upcount")
     @Operation(summary = "게시물을 추천 높은순으로 정렬해 나열합니다.")
-    public List<PostReadDTO> findByUpCount(){
+    public List<PostReadDTO> findByUpCount() {
         return postService.findByUpCountPost();
     }
 
     @PatchMapping("/update")
     @Operation(summary = "게시물을 수정합니다.", description = "이거 실행하기전에 /post/decreaseUpCount에 가서 파일 첨부먼저 하고 리턴값을" +
             "fileName에 넣어줘야합니다.")
-    public ResponseEntity<?> updatePost(@RequestBody PostUpdateDTO postUpdateDTO, @RequestParam("postId") Long postId) {
+    public ResponseEntity<?> updatePost(
+            @RequestBody PostUpdateDTO postUpdateDTO,
+            @RequestParam("postId") Long postId,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한확인 - 글 작성자만 가능
+        if (userId == null || !userId.equals(postService.getWriterUserId(postId)))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         return postService.updatePost(postId, postUpdateDTO);
     }
 
     @DeleteMapping("/delete")
     @Operation(summary = "게시물을 삭제합니다.")
-    public ResponseEntity<?> deletePost(@RequestParam("postId") Long postId) {
+    public ResponseEntity<?> deletePost(
+            @RequestParam("postId") Long postId,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한확인 - 글 작성자만 가능
+        if (userId == null || !userId.equals(postService.getWriterUserId(postId)))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         return postService.deletePost(postId);
     }
 
 
     @PostMapping("/increase/UpCount")
-    @Operation(summary = "게시물 채택" , description = "postid : 채택한 게시물 id, state 추천 상태 true : 채택 / 미채택이면 그냥 없어짐")
-    public ResponseEntity<?> increaeUpCount(@RequestParam("postId") Long postId, @RequestParam("userId") Long userId) {
-        return postService.IncreaseUpCountPost(postId,userId);
+    @Operation(summary = "게시물 채택", description = "postid : 채택한 게시물 id, state 추천 상태 true : 채택 / 미채택이면 그냥 없어짐")
+    public ResponseEntity<?> increaeUpCount(
+            @RequestParam("postId") Long postId,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한확인 - 로그인 되어있으면 가능
+        if (userId == null)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        return postService.IncreaseUpCountPost(postId, userId);
     }
 
     @PostMapping("/decrease/UpCount")
     @Operation(summary = "게시물 채택 취소", description = "postid : 채택한 게시물 id, state 추천 상태 true : 채택 / 미채택이면 그냥 없어짐")
-    public ResponseEntity<?> decreaseUpCount(@RequestParam("postId") Long postId, @RequestParam("userId") Long userId) {
-        return postService.decreaseUpCountPost(postId,userId);
+    public ResponseEntity<?> decreaseUpCount(
+            @RequestParam("postId") Long postId,
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        // 권한확인 - 로그인 되어있으면 가능
+        if (userId == null)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        return postService.decreaseUpCountPost(postId, userId);
     }
 
 }
